Refactoring assets
