+++ Dockerfile_app.REPLACE
@@ -38,14 +38,15 @@ RUN apt-get update && apt-get install -y \
 RUN sh -c "$(wget -O- https://github.com/deluan/zsh-in-docker/releases/download/v1.1.1/zsh-in-docker.sh)"
 
 RUN pip3 install \
-    dagster==1.1.3 \
-    dagit==1.1.3 \
-    dagster-graphql==1.1.3 \
-    dagster-postgres==0.17.3 \
-    dagster-docker==0.17.3 \
-    dagster_shell==0.17.3 \
-    dagster_snowflake==0.17.3 \
-    dagster_dbt==0.17.3 \
+    dagster==1.1.6 \
+    dagit==1.1.6 \
+    dagster-graphql==1.1.6 \
+    dagster-postgres==0.17.6 \
+    dagster-docker==0.17.6 \
+    dagster_shell==0.17.6 \
+    dagster_snowflake==0.17.6 \
+    dagster_dbt==0.17.6 \
+    dagster_pandas==0.17.6 \
     dagster_hex \
     dbt-core==1.3.1 \
     dbt-snowflake==1.3.0 \
