+++ Dockerfile_dagster
@@ -8,12 +8,12 @@ ENV DAGSTER_HOME=/opt/dagster/dagster_home/
 
 
 RUN pip install \
-    dagster==1.1.3 \
-    dagit==1.1.3 \
-    dagster-graphql==1.1.3 \
-    dagster-postgres==0.17.3 \
-    dagster-docker==0.17.3 \
-    dagster_shell==0.17.3
+    dagster==1.1.6 \
+    dagit==1.1.6 \
+    dagster-graphql==1.1.6 \
+    dagster-postgres==0.17.6 \
+    dagster-docker==0.17.6 \
+    dagster_shell==0.17.6
 
 # Copy dagster instance and workspace YAML
 RUN mkdir -p $DAGSTER_HOME
