+++ discursus_data_platform/assets/data_apps/public_dashboard_assets.py
@@ -42,37 +42,3 @@ def hex_main_dashboard_refresh(context):
             ),
         },
     )
-
-
-@asset(
-    description = "Hex daily assets refresh",
-    group_name = "data_apps",
-    resource_defs = {
-        'hex_resource': resources.my_resources.my_hex_resource
-    },
-)
-def hex_daily_assets_refresh(context):
-    hex_output: HexOutput = context.resources.hex_resource.run_and_poll(
-        project_id = "fa81df75-74df-4bb7-b7fe-9920dc00d99e",
-        inputs = None,
-        update_cache = True,
-        kill_on_timeout = True,
-        poll_interval = DEFAULT_POLL_INTERVAL,
-        poll_timeout = None,
-    )
-    asset_name = ["hex", hex_output.run_response["projectId"]]
-
-    return Output(
-        value = asset_name, 
-        metadata = {
-            "run_url": MetadataValue.url(hex_output.run_response["runUrl"]),
-            "run_status_url": MetadataValue.url(
-                hex_output.run_response["runStatusUrl"]
-            ),
-            "trace_id": MetadataValue.text(hex_output.run_response["traceId"]),
-            "run_id": MetadataValue.text(hex_output.run_response["runId"]),
-            "elapsed_time": MetadataValue.int(
-                hex_output.status_response["elapsedTime"]
-            ),
-        },
-    )
