+++ discursus_data_platform/assets/sources/gdelt_source_assets.py
@@ -1,0 +1,83 @@
+from dagster import asset, Output
+from dagster_pandas import DataFrame
+import pandas as pd
+import boto3
+from io import StringIO
+import resources.my_resources
+from resources.ml_enrichment_tracker import MLEnrichmentJobTracker
+
+
+@asset(
+    description = "List of events mined on GDELT",
+    group_name = "sources",
+    resource_defs = {
+        'aws_resource': resources.my_resources.my_aws_resource,
+        'gdelt_resource': resources.my_resources.my_gdelt_resource,
+        'snowflake_resource': resources.my_resources.my_snowflake_resource
+    }
+)
+def gdelt_events(context) -> DataFrame:
+    # Build source path
+    latest_events_url = context.resources.gdelt_resource.get_url_to_latest_asset("events")
+    gdelt_asset_filename_zip = str(latest_events_url).split('gdeltv2/')[1]
+    gdelt_asset_filename_csv = gdelt_asset_filename_zip.split('.zip')[0]
+    gdelt_asset_filedate = gdelt_asset_filename_csv[0:8]
+    gdelt_asset_source_path = 'sources/gdelt/' + gdelt_asset_filedate + '/' + gdelt_asset_filename_csv
+
+    # Mine and filter data
+    df_latest_events = context.resources.gdelt_resource.mine_latest_asset(latest_events_url)
+    df_latest_events_filtered = context.resources.gdelt_resource.filter_latest_events(df_latest_events, 14, ['US', 'CA'])
+    
+    # Save data to S3
+    context.resources.aws_resource.s3_put(df_latest_events_filtered, 'discursus-io', gdelt_asset_source_path)
+
+    # Transfer to Snowflake
+    q_load_gdelt_events = "alter pipe gdelt_events_pipe refresh;"
+    context.resources.snowflake_resource.execute_query(q_load_gdelt_events)
+
+    # Return asset
+    return Output(
+        value = df_latest_events_filtered, 
+        metadata = {
+            "path": "s3://discursus-io/" + gdelt_asset_source_path,
+            "rows": df_latest_events_filtered.index.size
+        }
+    )
+
+
+@asset(
+    description = "List of mentions mined from GDELT",
+    group_name = "sources",
+    resource_defs = {
+        'aws_resource': resources.my_resources.my_aws_resource,
+        'gdelt_resource': resources.my_resources.my_gdelt_resource,
+        'snowflake_resource': resources.my_resources.my_snowflake_resource
+    }
+)
+def gdelt_mentions(context, gdelt_events) -> DataFrame:
+    # Build source path
+    latest_mentions_url = context.resources.gdelt_resource.get_url_to_latest_asset("mentions")
+    gdelt_asset_filename_zip = str(latest_mentions_url).split('gdeltv2/')[1]
+    gdelt_asset_filename_csv = gdelt_asset_filename_zip.split('.zip')[0]
+    gdelt_asset_filedate = gdelt_asset_filename_csv[0:8]
+    gdelt_asset_source_path = 'sources/gdelt/' + gdelt_asset_filedate + '/' + gdelt_asset_filename_csv
+
+    # Mine and filter data
+    df_latest_mentions = context.resources.gdelt_resource.mine_latest_asset(latest_mentions_url)
+    df_latest_mentions_filtered = context.resources.gdelt_resource.filter_latest_mentions(df_latest_mentions, gdelt_events)
+    
+    # Save data to S3
+    context.resources.aws_resource.s3_put(df_latest_mentions_filtered, 'discursus-io', gdelt_asset_source_path)
+
+    # Transfer to Snowflake
+    q_load_gdelt_mentions_events = "alter pipe gdelt_mentions_pipe refresh;"
+    context.resources.snowflake_resource.execute_query(q_load_gdelt_mentions_events)
+
+    # Return asset
+    return Output(
+        value = df_latest_mentions_filtered, 
+        metadata = {
+            "path": "s3://discursus-io/" + gdelt_asset_source_path,
+            "rows": df_latest_mentions_filtered.index.size
+        }
+    )
